Install Locale::PO within the recipe
This is cleaner as it means the install happens under the temporary
HOME created by homebrew, and means it actually works.  It also
means one less manual step.  The slight downside is that Locale::PO
has to be fetched and installed each time survex --HEAD is, but
that isn't a big deal.